requirements.txt
bioformats
pyfaidx
pyvcf
